[policies] refactor pack, wrapper and Root components
