+++ app/javascript/src/Policies/components/PoliciesWrapper.tsx
@@ -1,12 +1,12 @@
-import Root from 'Policies/components/Root'
+import { Provider } from 'react-redux'
+
+import PoliciesWidget from 'Policies/components/PoliciesWidget'
 import configureStore from 'Policies/store/configureStore'
 import { initialState } from 'Policies/reducers/initialState'
 import { populateChainFromConfigs } from 'Policies/actions'
 import { createReactWrapper } from 'utilities/createReactWrapper'
 
 import type { Action, PolicyConfig, RegistryPolicy } from 'Policies/types'
-
-import 'Policies/styles/policies.scss'
 
 interface Props {
   registry: RegistryPolicy[];
@@ -18,7 +18,12 @@ const PoliciesWrapper = ({ registry, chain, serviceId }: Props, elementId: strin
   const store = configureStore(initialState)
   store.dispatch(populateChainFromConfigs(serviceId, chain, registry) as unknown as Action)
 
-  createReactWrapper(<Root store={store} />, elementId)
+  createReactWrapper(
+    <Provider store={store}>
+      <PoliciesWidget />
+    </Provider>,
+    elementId
+  )
 }
 
 export type { Props }
