--- spec/javascripts/Policies/components/Root.spec.tsx
@@ -1,1 +1,0 @@
-it.todo('test Root')
